+++ src/components/WalletUnlock/NewTrezor.tsx
@@ -5,7 +5,7 @@ import prop from 'ramda/src/prop';
 import { MOONPAY_ASSET_UUIDS } from '@utils';
 import { FormData, WalletId, ExtendedAsset } from '@types';
 import translate, { translateRaw } from '@translations';
-import { Spinner, Button, DeterministicAccountList, AssetDropdown } from '@components';
+import { Spinner, Button, DeterministicAccountList, AssetSelector } from '@components';
 import {
   EXT_URLS,
   TREZOR_DERIVATION_PATHS,
@@ -68,7 +68,7 @@ const TrezorDecrypt = ({ formData, onUnlock }: OwnProps) => {
   if (state.isConnected && state.asset && (state.queuedAccounts || state.finishedAccounts)) {
     return (
       <div className="Mnemonic-dpath">
-        <AssetDropdown
+        <AssetSelector
           selectedAsset={assetToUse}
           assets={filteredAssets}
           onSelect={(option: ExtendedAsset) => {
