+++ src/components/index.ts
@@ -4,9 +4,8 @@ export { default as AccountDropdown } from './AccountDropdown';
 export { default as AccountSummary } from './AccountSummary';
 export { default as AccountList } from './AccountList';
 export { default as AmountInput } from './AmountInput';
-export { default as AssetDropdownItem } from './AssetDropdownItem';
 export { default as AssetIcon } from './AssetIcon';
-export { default as AssetDropdown } from './AssetDropdown';
+export { default as AssetSelector } from './AssetSelector';
 export { default as Button } from './Button';
 export { default as Checkbox } from './Checkbox';
 export { default as CopyableCodeBlock } from './CopyableCodeBlock';
